backend/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-cors-headers==3.10.0
tensorflow
joblib
scikit-learn